backend/requirements.txt
flask
flask-cors
flask-mysqldb
flask-bcrypt
pyjwt
python-dotenv